Add countAllMembers() with tests. Update tests to use try/catch

+++ dspace-api/src/test/java/org/dspace/eperson/EPersonTest.java
@@ -1041,36 +1041,46 @@ public class EPersonTest extends AbstractUnitTest {
         EPerson eperson3 = createEPersonAndAddToGroup("test3@example.com", group);
         groupService.update(context, group);
 
-        // Assert that findByGroup is the same list of EPersons as getMembers() when pagination is ignored
-        // (NOTE: Pagination is tested in GroupRestRepositoryIT)
-        // NOTE: isEqualCollection() must be used for comparison because Hibernate's "PersistentBag" cannot be compared
-        // directly to a List. See https://stackoverflow.com/a/57399383/3750035
-        assertTrue(CollectionUtils.isEqualCollection(group.getMembers(),
-                                                     ePersonService.findByGroups(context, Set.of(group), -1, -1)));
-        // Assert countByGroups is the same as the size of members
-        assertEquals(group.getMembers().size(), ePersonService.countByGroups(context, Set.of(group)));
+        Group group2 = null;
+        EPerson eperson4 = null;
 
-        // Add another group with duplicate EPerson
-        Group group2 = createGroup("anotherGroup");
-        groupService.addMember(context, group2, eperson1);
-        groupService.update(context, group2);
+        try {
+            // Assert that findByGroup is the same list of EPersons as getMembers() when pagination is ignored
+            // (NOTE: Pagination is tested in GroupRestRepositoryIT)
+            // NOTE: isEqualCollection() must be used for comparison because Hibernate's "PersistentBag" cannot be
+            // compared directly to a List. See https://stackoverflow.com/a/57399383/3750035
+            assertTrue(
+                CollectionUtils.isEqualCollection(group.getMembers(),
+                                                  ePersonService.findByGroups(context, Set.of(group), -1, -1)));
+            // Assert countByGroups is the same as the size of members
+            assertEquals(group.getMembers().size(), ePersonService.countByGroups(context, Set.of(group)));
 
-        // Verify countByGroups is still 3 (existing person should not be counted twice)
-        assertEquals(3, ePersonService.countByGroups(context, Set.of(group, group2)));
+            // Add another group with duplicate EPerson
+            group2 = createGroup("anotherGroup");
+            groupService.addMember(context, group2, eperson1);
+            groupService.update(context, group2);
 
-        // Add a new EPerson to new group, verify count goes up by one
-        EPerson eperson4 = createEPersonAndAddToGroup("test4@example.com", group2);
-        assertEquals(4, ePersonService.countByGroups(context, Set.of(group, group2)));
+            // Verify countByGroups is still 3 (existing person should not be counted twice)
+            assertEquals(3, ePersonService.countByGroups(context, Set.of(group, group2)));
 
-        // Clean up our data
-        context.turnOffAuthorisationSystem();
-        groupService.delete(context, group);
-        groupService.delete(context, group2);
-        ePersonService.delete(context, eperson1);
-        ePersonService.delete(context, eperson2);
-        ePersonService.delete(context, eperson3);
-        ePersonService.delete(context, eperson4);
-        context.restoreAuthSystemState();
+            // Add a new EPerson to new group, verify count goes up by one
+            eperson4 = createEPersonAndAddToGroup("test4@example.com", group2);
+            assertEquals(4, ePersonService.countByGroups(context, Set.of(group, group2)));
+        } finally {
+            // Clean up our data
+            context.turnOffAuthorisationSystem();
+            groupService.delete(context, group);
+            if (group2 != null) {
+                groupService.delete(context, group2);
+            }
+            ePersonService.delete(context, eperson1);
+            ePersonService.delete(context, eperson2);
+            ePersonService.delete(context, eperson3);
+            if (eperson4 != null) {
+                ePersonService.delete(context, eperson4);
+            }
+            context.restoreAuthSystemState();
+        }
     }
 
     /**
